requirements.txt
numpy
opencv-python
opencv-contrib-python
mysql-connector-python
pyttsx3
Pillow
pysimplegui
python-dotenv
PyQt5